+++ src/GameFinder.Launcher.Heroic/HeroicGOGHandler.cs
@@ -17,7 +17,7 @@ using OneOf;
 namespace GameCollector.Launcher.Heroic;
 
 [PublicAPI]
-public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
+public class HeroicGOGHandler : AHandler<HeroicGOGGame, GOGGameId>
 {
     private readonly IFileSystem _fileSystem;
     private readonly ILogger _logger;
@@ -39,7 +39,7 @@ public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
     }
 
     /// <inheritdoc/>
-    public override Func<GOGGame, GOGGameId> IdSelector { get; } = static game => game.Id;
+    public override Func<HeroicGOGGame, GOGGameId> IdSelector { get; } = static game => game.Id;
 
     /// <inheritdoc/>
     public override IEqualityComparer<GOGGameId>? IdEqualityComparer => null;
@@ -51,7 +51,7 @@ public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
     }
 
     /// <inheritdoc/>
-    public override IEnumerable<OneOf<GOGGame, ErrorMessage>> FindAllGames(Settings? settings = null)
+    public override IEnumerable<OneOf<HeroicGOGGame, ErrorMessage>> FindAllGames(Settings? settings = null)
     {
         var configDirectory = FindConfigDirectory(_fileSystem)
             .FirstOrDefault(path => path.DirectoryExists());
@@ -77,7 +77,8 @@ public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
         }
     }
 
-    internal IEnumerable<OneOf<GOGGame, ErrorMessage>> ParseInstalledJsonFile(AbsolutePath path, AbsolutePath configPath)
+    [RequiresUnreferencedCode("Calls System.Text.Json.JsonSerializer.Deserialize<TValue>(Stream, JsonSerializerOptions)")]
+    internal IEnumerable<OneOf<HeroicGOGGame, ErrorMessage>> ParseInstalledJsonFile(AbsolutePath path, AbsolutePath configPath)
     {
         using var stream = path.Open(FileMode.Open, FileAccess.Read, FileShare.Read);
         var root = JsonSerializer.Deserialize<DTOs.Root>(stream, JsonSerializerOptions);
@@ -89,7 +90,7 @@ public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
 
         foreach (var installed in root.Installed)
         {
-            OneOf<GOGGame, ErrorMessage> res;
+            OneOf<HeroicGOGGame, ErrorMessage> res;
             try
             {
                 res = Parse(installed, configPath, path.FileSystem);
@@ -103,7 +104,8 @@ public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
         }
     }
 
-    internal OneOf<GOGGame, ErrorMessage> Parse(
+    [RequiresUnreferencedCode("Calls GameFinder.Launcher.Heroic.HeroicGOGHandler.GetWineData(Installed, AbsolutePath, Int64)")]
+    internal OneOf<HeroicGOGGame, ErrorMessage> Parse(
         DTOs.Installed installed,
         AbsolutePath configPath,
         IFileSystem fileSystem)
@@ -111,6 +113,16 @@ public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
         if (!long.TryParse(installed.AppName, NumberStyles.Integer, CultureInfo.InvariantCulture, out var id))
         {
             return new ErrorMessage($"The value \"appName\" is not a number: \"{installed.AppName}\"");
+        }
+
+        ulong buildId;
+        if (string.IsNullOrWhiteSpace(installed.BuildId))
+        {
+            buildId = 0;
+        }
+        else if (!ulong.TryParse(installed.BuildId, CultureInfo.InvariantCulture, out buildId))
+        {
+            return new ErrorMessage($"The value \"buildID\" is not a number: \"{installed.BuildId}\"");
         }
 
         var installedPlatform = installed.Platform switch
@@ -130,8 +142,17 @@ public class HeroicGOGHandler : AHandler<GOGGame, GOGGameId>
             }
         }
 
+        var installedDLCs = new List<GOGGameId>();
+        foreach (var sRawId in installed.InstalledDLCs)
+        {
+            if (long.TryParse(sRawId, NumberStyles.Integer, CultureInfo.InvariantCulture, out var rawId))
+            {
+                installedDLCs.Add(GOGGameId.From(rawId));
+            }
+        }
+
         var path = fileSystem.FromUnsanitizedFullPath(installed.InstallPath);
-        return new HeroicGOGGame(GOGGameId.From(id), installed.AppName, path, installed.BuildId, wineData, installedPlatform);
+        return new HeroicGOGGame(GOGGameId.From(id), installed.AppName, path, buildId, wineData, installedPlatform, installedDLCs);
     }
 
     [RequiresUnreferencedCode("Calls System.Text.Json.JsonSerializer.Deserialize<TValue>(JsonSerializerOptions)")]