+++ src/GameFinder.StoreHandlers.GOG/GOGGame.cs
@@ -1,3 +1,4 @@
+using System;
 using GameFinder.Common;
 using JetBrains.Annotations;
 using NexusMods.Paths;
@@ -14,6 +15,7 @@ namespace GameCollector.StoreHandlers.GOG;
 /// <param name="Name"></param>
 /// <param name="Path"></param>
 /// <param name="BuildId"></param>
+/// <param name="ParentGameId"></param>
 /// <param name="Launch"></param>
 /// <param name="LaunchParam"></param>
 /// <param name="LaunchUrl"></param>
@@ -26,7 +28,6 @@ namespace GameCollector.StoreHandlers.GOG;
 /// <param name="IsHidden"></param>
 /// <param name="Tags"></param>
 /// <param name="MyRating"></param>
-/// <param name="ParentId"></param>
 /// <param name="ReleaseDate"></param>
 /// <param name="StoreUrl"></param>
 /// <param name="SupportUrl"></param>
@@ -38,6 +39,7 @@ public record GOGGame(GOGGameId Id,
                       string Name,
                       AbsolutePath Path,
                       ulong BuildId,
+                      Nullable<GOGGameId> ParentGameId = null,
                       AbsolutePath Launch = new(),
                       string LaunchParam = "",
                       string LaunchUrl = "",
@@ -50,7 +52,6 @@ public record GOGGame(GOGGameId Id,
                       bool IsHidden = false,
                       IList<string>? Tags = null,
                       ushort? MyRating = null,
-                      GOGGameId? ParentId = default,
                       DateTime? ReleaseDate = null,
                       string StoreUrl = "",
                       string SupportUrl = "",
@@ -73,7 +74,7 @@ public record GOGGame(GOGGameId Id,
              IsHidden: IsHidden,
              Tags: Tags,
              MyRating: MyRating,
-             BaseGame: ParentId.ToString(),
+             BaseGame: ParentGameId.ToString(),
              Metadata: new Dictionary<string, List<string>>(StringComparer.OrdinalIgnoreCase)
              {
                  ["ReleaseDate"] = new() { ReleaseDate is null ? "" : ((DateTime)ReleaseDate).ToString(CultureInfo.InvariantCulture) },