+++ tests/GameFinder.StoreHandlers.GOG.Tests/Test_ShouldError_InvalidValues.cs
@@ -15,6 +15,10 @@ public partial class GOGTests
         var invalidKey = gogKey.AddSubKey(keyName);
         invalidKey.AddValue("gameId", gameId);
 
+        /*
+        var error = handler.ShouldOnlyBeOneError();
+        error.Message.Should().EndWith($"The value \"gameID\" of {invalidKey.GetName()} is not a number: \"{gameId}\"");
+        */
         foreach (var error in handler.ShouldOnlyBeErrors())
         {
             error.Should().BeOneOf($"The value \"gameID\" of {invalidKey.GetName()} is not a number: \"{gameId}\"", "GOG database not found.");